Warn when signatures are valid but untrusted
When verifying a signature seahorse-tool currently behave just the same if the
signing key is trusted or not. Given that the only the uid of the key is given
in the notification, this opens the door to some attacks.

Display two different messages depending on the validity
level of the signing key.

https://bugzilla.gnome.org/show_bug.cgi?id=708640